Review changes: move to another file, add tests, add behind feature
