+++ src/alloc/mod.rs
@@ -2,11 +2,9 @@
 
 #![cfg(feature = "alloc")]
 
+pub mod allocext;
+
 extern crate std;
-
-#[cfg(target_os = "linux")]
-use self::memfd_secret_alloc::*;
-
 use self::raw_alloc::*;
 use self::std::process::abort;
 use self::std::sync::Once;
@@ -63,42 +61,6 @@ mod raw_alloc {
     pub unsafe fn free_aligned(memptr: *mut u8, size: usize) {
         let layout = Layout::from_size_align_unchecked(size, PAGE_SIZE);
         dealloc(memptr, layout);
-    }
-}
-
-#[cfg(target_os = "linux")]
-mod memfd_secret_alloc {
-    use core::convert::TryInto;
-
-    use super::*;
-
-    #[inline]
-    pub unsafe fn alloc_memfd_secret(size: usize) -> Option<NonNull<u8>> {
-        let fd: Result<libc::c_int, _> = libc::syscall(libc::SYS_memfd_secret, 0).try_into();
-
-        if fd.is_err() || fd.unwrap() < 0 {
-            return None;
-        }
-
-        let fd = fd.unwrap();
-
-        // File size is set using ftruncate
-        let _ = libc::ftruncate(fd, size as libc::off_t);
-
-        let ptr = libc::mmap(
-            ptr::null_mut(),
-            size,
-            Prot::ReadWrite,
-            libc::MAP_SHARED,
-            fd,
-            0,
-        );
-
-        if ptr == libc::MAP_FAILED {
-            return None;
-        }
-
-        NonNull::new(ptr as *mut u8)
     }
 }
 
@@ -219,41 +181,6 @@ unsafe fn _malloc(size: usize) -> Option<*mut u8> {
     Some(user_ptr)
 }
 
-#[cfg(target_os = "linux")]
-unsafe fn _memfd_secret(size: usize) -> Option<*mut u8> {
-    ALLOC_INIT.call_once(|| alloc_init());
-
-    if size >= ::core::usize::MAX - PAGE_SIZE * 4 {
-        return None;
-    }
-
-    // aligned alloc ptr
-    let size_with_canary = CANARY_SIZE + size;
-    let unprotected_size = page_round(size_with_canary);
-    let total_size = PAGE_SIZE + PAGE_SIZE + unprotected_size + PAGE_SIZE;
-    let base_ptr = alloc_memfd_secret(total_size)?.as_ptr();
-    let unprotected_ptr = base_ptr.add(PAGE_SIZE * 2);
-
-    // mprotect can be used to change protection flag after mmap setup
-    // https://www.gnu.org/software/libc/manual/html_node/Memory-Protection.html#index-mprotect
-    _mprotect(base_ptr.add(PAGE_SIZE), PAGE_SIZE, Prot::NoAccess);
-    _mprotect(
-        unprotected_ptr.add(unprotected_size),
-        PAGE_SIZE,
-        Prot::NoAccess,
-    );
-
-    let canary_ptr = unprotected_ptr.add(unprotected_size - size_with_canary);
-    let user_ptr = canary_ptr.add(CANARY_SIZE);
-    ptr::copy_nonoverlapping(CANARY.as_ptr(), canary_ptr, CANARY_SIZE);
-    ptr::write_unaligned(base_ptr as *mut usize, unprotected_size);
-    _mprotect(base_ptr, PAGE_SIZE, Prot::ReadOnly);
-
-    assert_eq!(unprotected_ptr_from_user_ptr(user_ptr), unprotected_ptr);
-
-    Some(user_ptr)
-}
-
 /// Secure `malloc`.
 #[inline]
 pub unsafe fn malloc<T>() -> Option<NonNull<T>> {
@@ -267,25 +194,6 @@ pub unsafe fn malloc<T>() -> Option<NonNull<T>> {
 #[inline]
 pub unsafe fn malloc_sized(size: usize) -> Option<NonNull<[u8]>> {
     _malloc(size).map(|memptr| {
-        ptr::write_bytes(memptr, GARBAGE_VALUE, size);
-        NonNull::new_unchecked(slice::from_raw_parts_mut(memptr, size))
-    })
-}
-
-#[inline]
-#[cfg(target_os = "linux")]
-pub unsafe fn memfd_secret<T>() -> Option<NonNull<T>> {
-    _memfd_secret(mem::size_of::<T>()).map(|memptr| {
-        ptr::write_bytes(memptr, GARBAGE_VALUE, mem::size_of::<T>());
-        NonNull::new_unchecked(memptr as *mut T)
-    })
-}
-
-/// Secure `malloc_sized`.
-#[inline]
-#[cfg(target_os = "linux")]
-pub unsafe fn memfd_secret_sized(size: usize) -> Option<NonNull<[u8]>> {
-    _memfd_secret(size).map(|memptr| {
         ptr::write_bytes(memptr, GARBAGE_VALUE, size);
         NonNull::new_unchecked(slice::from_raw_parts_mut(memptr, size))
     })
@@ -313,34 +221,4 @@ pub unsafe fn free<T: ?Sized>(memptr: NonNull<T>) {
     crate::munlock(unprotected_ptr, unprotected_size);
 
     free_aligned(base_ptr, total_size);
-}
-
-/// Secure `free` for memfd_secret,
-/// i.e. provides read write access back to mprotect guard pages
-/// and unmaps mmap
-#[cfg(target_os = "linux")]
-pub unsafe fn free_memfd_secret<T: ?Sized>(memptr: NonNull<T>) {
-    use libc::c_void;
-
-    let memptr = memptr.as_ptr() as *mut u8;
-
-    // get unprotected ptr
-    let canary_ptr = memptr.sub(CANARY_SIZE);
-    let unprotected_ptr = unprotected_ptr_from_user_ptr(memptr);
-    let base_ptr = unprotected_ptr.sub(PAGE_SIZE * 2);
-    let unprotected_size = ptr::read(base_ptr as *const usize);
-
-    // check
-    if !crate::memeq(canary_ptr as *const u8, CANARY.as_ptr(), CANARY_SIZE) {
-        abort();
-    }
-
-    // free
-    let total_size = PAGE_SIZE + PAGE_SIZE + unprotected_size + PAGE_SIZE;
-    _mprotect(base_ptr, total_size, Prot::ReadWrite);
-
-    let res = libc::munmap(base_ptr as *mut c_void, total_size);
-    if res < 0 {
-        abort();
-    }
 }
