+++ src/lib.rs
@@ -13,9 +13,10 @@ pub use mlock::{mlock, munlock};
 #[cfg(feature = "alloc")]
 pub use alloc::{free, malloc, malloc_sized, mprotect, Prot};
 
-#[cfg(feature = "alloc")]
+#[cfg(feature = "alloc_ext")]
 #[cfg(target_os = "linux")]
-pub use alloc::{free_memfd_secret, memfd_secret, memfd_secret_sized};
+pub use alloc::allocext::{free_memfd_secret, memfd_secret, memfd_secret_sized};
+
 // -- memcmp --
 
 /// Secure `memeq`.
